apps/core-api/pyproject.toml
[project]
name = "signet-core-api"
version = "0.1.0"
description = "Signet Protocol Core API (FastAPI)"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.35.0",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.10.0",
  "orjson>=3.10.6",
  "python-multipart>=0.0.9",
  "PyNaCl>=1.5.0",
  "rfc8785>=0.1.0",
  "prometheus-client>=0.22.1",
  "httpx>=0.27.2",
  "structlog>=24.1.0",
  "opentelemetry-api>=1.26.0",
  "opentelemetry-sdk>=1.26.0",
  "opentelemetry-instrumentation-fastapi>=0.47b0"
]

[project.optional-dependencies]
dev = [
  "uvicorn[standard]>=0.35.0",
  "pytest>=8.3",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "ruff>=0.5",
  "mypy>=1.10",
  "bandit>=1.7",
  "pip-audit>=2.7",
  "pydantic>=2.8",
  "fastapi>=0.115"
]

[tool.pytest.ini_options]
addopts = "-q --disable-warnings --maxfail=1"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "I", "UP", "B", "C4", "SIM", "N", "S", "RUF"]
format.preview = true

[tool.mypy]
python_version = "3.12"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["server"]
